+++ CHANGELOG.md
@@ -34,3 +34,4 @@
 - Added terms of service state endpoint
 - Added terms of service state endpoint
 - Added account online return policy endpoint
+- Added account homepage endpoint